chore: fix issues
