+++ packages/native-ui/src/components/FormField/FormFielInvalidIcon.tsx
@@ -4,8 +4,8 @@ import IconProps from '../Icon/Icon.props';
 import { SvgRef } from '../../types';
 import { WarningMediumContainedIcon } from '@utilitywarehouse/react-native-icons';
 
-const FormFieldInvalidIcon = forwardRef<SvgRef, Omit<IconProps, 'as'>>(props => {
-  return <HelperIcon as={WarningMediumContainedIcon} {...props} />;
+const FormFieldInvalidIcon = forwardRef<SvgRef, Omit<IconProps, 'as'>>((props, ref) => {
+  return <HelperIcon ref={ref} as={WarningMediumContainedIcon} {...props} />;
 });
 
 FormFieldInvalidIcon.displayName = 'FormFieldInvalidIcon';
